requirements.txt
Flask==2.1.1
huggingface-hub==0.15.1
pydantic==1.10.2
python-dotenv==0.21.0
